Floral/requirements.txt
Django==2.0.5
djangorestframework==3.8.2
Pygments==2.2.0
pytz==2018.4
